clontech/clontech-portal/requirements.txt
amqp==2.5.2
arrow==0.12.1
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==1.11.29
django-appconf==1.0.4
django-celery-email==2.0.2
django-countries==6.1.2
django-debug-toolbar==2.0
django-debug-toolbar-template-profiler==1.0.2
django-extensions==2.1.9
django-extra-views==0.12.0
django-oauth-toolkit==1.1.3
django-pipeline==1.7.0
django-rest-framework==0.1.0
django-sendfile==0.3.11
django-tz-detect==0.2.10
djangorestframework==3.11.0
djangorestframework-camel-case==0.2.0
djangorestframework-csv==2.1.0
envdir==1.0.1
Faker==4.0.2
freezegun==0.3.15
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
markdown2==2.3.8
mysqlclient==1.3.14
ndx-app-versions==5.5.12
ndx-audit-log==2.2.0
ndx-auth-source==5.5.12
ndx-django-utils==5.5.12
ndx-location==5.5.12
ndx-mobile-device-support==5.5.12
ndx-mpaf-source==5.5.12
ndx-organisation==5.5.12
ndx-portal-source==5.5.12
ndx-result-source==5.5.12
novarum-textutils==2.1
oauthlib==3.1.0
pdfrw==0.4
Pillow==7.0.0
pycparser==2.20
pylibmc==1.6.1
Pyphen==0.9.5
pyStrich==0.8
python-dateutil==2.8.1
pytz==2019.3
raven==6.10.0
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
tinycss2==1.0.2
typing==3.7.4.1
unicodecsv==0.14.1
unittest-xml-reporting==3.0.2
urllib3==1.25.8
vine==1.3.0
WeasyPrint==0.42.3
webencodings==0.5.1
zipp==1.2.0
